final-pjt-back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
idna==3.7
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
